achieved stability in the retrieval of large data that can comply to rate limits set by the receiving api at data.gov
key changes made

1. flatMap is now controlled in concurrency via CONCURRENCYLIMIT parameter that can be statically adjusted to avoid rate limiting

2. the large async operations have been serialized to reduce sudden load on the receiving api. pritoritizing the BaseAPi initially and ExposInfo coming after.

these changes were made in APICollationService.java

cacheable has been tested about due to the error with key= attribute. default seemed to work well but SpEL named parameter #address did not seem to work. reason was the compiler needing to be run with additional -parameters option but this causes trouble with JPQL queries being made in repository.

positional arguments #p0 , ... , #pn has been used to address this concern.

------------ summary of the key changes in making async ----------

APICollationService.java

1. the completableFutures serves both concurrency and asynchronous procession of the methods

-> concurrency(independent tasks to retrieve api from diff places or parallelism if the whole operation to preget is considered a larger foci) is achieved via the use of CompletableFuture -> refer to the sideEffects .

-> async is achieved via the SupplyAsync -> runAsync is also more appropriate here as the return values are not used, but supplyasync has been chosen for the reason to debug, as it can capture exceptions that bubble out through the try catch block. this allows the code to run with a headstart right after the baseApi collation finishes to pre-fetch the related data from the data.gov website

2. the issue about mutex(mutual exclusion) has been addressed

-> situation where mutex arises -> should user select a baseApi item triggering calls for exposInfoItem before the ExposInfoItem preget finishes -> this will cause duplicate api calls to start , as the cache put into redis has yet been conducted. this is mitigated through the usage of ReentrantLock -> implements Lock.class() , that locks the initial call based on the address,     ReentrantLock fetchLock = addressLocks.computeIfAbsent(address, k -> new ReentrantLock()); -> where the addresses being used are stored in an instance variable     private final ConcurrentHashMap<Address, ReentrantLock> addressLocks = new ConcurrentHashMap<>(); where upon completion the data is removed from the concurrentHashmap to save memory, before the locks unlock.

to account for the processes locked while the process finishes, additional check to the cache is present inside the lock where if the cache returns non -null it will simply cause the thread to return the stored value in the cache (cache hit) that has been filled by the first entering thread to the method

3. accounting for rate limit
the rate limit is accounted for via the flatmap's second parameter that limits the total number of operations that flatmap can do, that is statically controlled -> magic value -> within the WebClientApiService.java

this details all the key changes made.